dynamic host volumes: send register to client for fingerprint (#24802)
When we register a volume without a plugin, we need to send a client RPC so that
the node fingerprint can be updated. The registered volume also needs to be
written to client state so that we can restore the fingerprint after a restart.

+++ client/hostvolumemanager/host_volumes.go
@@ -6,6 +6,8 @@ package hostvolumemanager
 import (
 	"context"
 	"errors"
+	"fmt"
+	"os"
 	"path/filepath"
 	"sync"
 
@@ -101,6 +103,7 @@ func (hvm *HostVolumeManager) Create(ctx context.Context,
 
 	volState := &cstructs.HostVolumeState{
 		ID:        req.ID,
+		HostPath:  pluginResp.Path,
 		CreateReq: req,
 	}
 	if err := hvm.stateMgr.PutDynamicHostVolume(volState); err != nil {
@@ -120,10 +123,10 @@ func (hvm *HostVolumeManager) Create(ctx context.Context,
 		}
 		// free up the volume name whether delete succeeded or not.
 		hvm.locker.release(req.Name)
-		return nil, helper.FlattenMultierror(err)
+		return nil, err
 	}
 
-	hvm.updateNodeVols(req.Name, genVolConfig(req, pluginResp))
+	hvm.updateNodeVols(req.Name, genVolConfig(req, pluginResp.Path))
 
 	resp := &cstructs.ClientHostVolumeCreateResponse{
 		VolumeName:    req.Name,
@@ -135,19 +138,59 @@ func (hvm *HostVolumeManager) Create(ctx context.Context,
 	return resp, nil
 }
 
+// Register saves the request to state, and updates the node with the volume.
+func (hvm *HostVolumeManager) Register(ctx context.Context,
+	req *cstructs.ClientHostVolumeRegisterRequest) error {
+
+	// can't have two of the same volume name w/ different IDs per client node
+	if err := hvm.locker.lock(req.Name, req.ID); err != nil {
+		return err
+	}
+
+	_, err := os.Stat(req.HostPath)
+	if err != nil {
+		hvm.locker.release(req.Name)
+		return fmt.Errorf("could not verify host path for %q: %w", req.Name, err)
+	}
+
+	// generate a stub create request and plugin response for the fingerprint
+	// and client state
+	creq := &cstructs.ClientHostVolumeCreateRequest{
+		ID:         req.ID,
+		Name:       req.Name,
+		NodeID:     req.NodeID,
+		Parameters: req.Parameters,
+	}
+	volState := &cstructs.HostVolumeState{
+		ID:        req.ID,
+		CreateReq: creq,
+		HostPath:  req.HostPath,
+	}
+	if err := hvm.stateMgr.PutDynamicHostVolume(volState); err != nil {
+		hvm.log.Error("failed to save volume in state", "volume_id", req.ID, "error", err)
+		hvm.locker.release(req.Name)
+		return err
+	}
+
+	hvm.updateNodeVols(req.Name, genVolConfig(creq, req.HostPath))
+	return nil
+}
+
 // Delete runs the appropriate plugin for the given request, removes it from
 // state, and updates the node to remove the volume.
 func (hvm *HostVolumeManager) Delete(ctx context.Context,
 	req *cstructs.ClientHostVolumeDeleteRequest) (*cstructs.ClientHostVolumeDeleteResponse, error) {
 
-	plug, err := hvm.getPlugin(req.PluginID)
-	if err != nil {
-		return nil, err
-	}
+	if req.PluginID != "" {
+		plug, err := hvm.getPlugin(req.PluginID)
+		if err != nil {
+			return nil, err
+		}
 
-	err = plug.Delete(ctx, req)
-	if err != nil {
-		return nil, err
+		err = plug.Delete(ctx, req)
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	if err := hvm.stateMgr.DeleteDynamicHostVolume(req.ID); err != nil {
@@ -193,41 +236,22 @@ func (hvm *HostVolumeManager) restoreFromState(ctx context.Context) (VolumeMap, 
 		return volumes, nil // nothing to do
 	}
 
-	// re-"create" the volumes - plugins have the best knowledge of their
-	// side effects, and they must be idempotent.
 	group := multierror.Group{}
 	for _, vol := range vols {
 		group.Go(func() error {
-			// missing plugins with associated volumes in state are considered
-			// client-stopping errors. they need to be fixed by cluster admins.
-			plug, err := hvm.getPlugin(vol.CreateReq.PluginID)
+			var volCfg *structs.ClientHostVolumeConfig
+			var err error
+			if vol.CreateReq.PluginID == "" {
+				volCfg, err = hvm.restoreForRegister(vol)
+			} else {
+				volCfg, err = hvm.restoreForCreate(ctx, vol)
+			}
 			if err != nil {
 				return err
 			}
 
-			// lock the name so future creates can't produce duplicates.
-			err = hvm.locker.lock(vol.CreateReq.Name, vol.CreateReq.ID)
-			// state should never have duplicate vol names, and restore happens
-			// prior to node registration, so new creates shouldn't come in
-			// concurrently, but check for error just in case.
-			if err != nil {
-				hvm.log.Error("error during restore",
-					"volume_name", vol.CreateReq.Name,
-					"volume_id", vol.CreateReq.ID,
-					"error", err)
-				// don't stop the world if it does happen, because an admin
-				// couldn't do anything about it short of wiping client state.
-				return nil
-			}
-
-			resp, err := plug.Create(ctx, vol.CreateReq)
-			if err != nil {
-				// plugin execution errors are only logged
-				hvm.log.Error("failed to restore", "plugin_id", vol.CreateReq.PluginID, "volume_id", vol.ID, "error", err)
-				return nil
-			}
 			mut.Lock()
-			volumes[vol.CreateReq.Name] = genVolConfig(vol.CreateReq, resp)
+			volumes[vol.CreateReq.Name] = volCfg
 			mut.Unlock()
 			return nil
 		})
@@ -236,13 +260,73 @@ func (hvm *HostVolumeManager) restoreFromState(ctx context.Context) (VolumeMap, 
 	return volumes, helper.FlattenMultierror(mErr.ErrorOrNil())
 }
 
+// restoreForCreate restores a single volume that was previously created by
+// Create, by "recreating" the volumes. Plugins have the best knowledge of their
+// side effects, and they must be idempotent.
+func (hvm *HostVolumeManager) restoreForCreate(ctx context.Context, vol *cstructs.HostVolumeState) (*structs.ClientHostVolumeConfig, error) {
+	// missing plugins with associated volumes in state are considered
+	// client-stopping errors. they need to be fixed by cluster admins.
+	plug, err := hvm.getPlugin(vol.CreateReq.PluginID)
+	if err != nil {
+		return nil, err
+	}
+
+	// lock the name so future creates can't produce duplicates.
+	err = hvm.locker.lock(vol.CreateReq.Name, vol.CreateReq.ID)
+	// state should never have duplicate vol names, and restore happens
+	// prior to node registration, so new creates shouldn't come in
+	// concurrently, but check for error just in case.
+	if err != nil {
+		hvm.log.Error("error during restore",
+			"volume_name", vol.CreateReq.Name,
+			"volume_id", vol.CreateReq.ID,
+			"error", err)
+		// don't stop the world if it does happen, because an admin
+		// couldn't do anything about it short of wiping client state.
+		return nil, nil
+	}
+
+	resp, err := plug.Create(ctx, vol.CreateReq)
+	if err != nil {
+		// plugin execution errors are only logged
+		hvm.log.Error("failed to restore",
+			"plugin_id", vol.CreateReq.PluginID, "volume_id", vol.ID, "error", err)
+		return nil, nil
+	}
+
+	return genVolConfig(vol.CreateReq, resp.Path), nil
+}
+
+// restoreForRegister restores a single volume that was previously created by
+// Register, by converting the stored struct. It otherwise behaves the same as
+// restoreForCreate.
+func (hvm *HostVolumeManager) restoreForRegister(vol *cstructs.HostVolumeState) (*structs.ClientHostVolumeConfig, error) {
+	err := hvm.locker.lock(vol.CreateReq.Name, vol.CreateReq.ID)
+	if err != nil {
+		hvm.log.Error("error during restore",
+			"volume_name", vol.CreateReq.Name,
+			"volume_id", vol.CreateReq.ID,
+			"error", err)
+		return nil, nil
+	}
+
+	_, err = os.Stat(vol.HostPath)
+	if err != nil {
+		hvm.log.Error("failed to restore: could not verify host path",
+			"volume_id", vol.ID, "error", err, "path", vol.HostPath)
+		return nil, nil
+	}
+
+	return genVolConfig(vol.CreateReq, vol.HostPath), nil
+}
+
 // genVolConfig generates the host volume config for the node to report as
 // available to the servers for job scheduling.
-func genVolConfig(req *cstructs.ClientHostVolumeCreateRequest, resp *HostVolumePluginCreateResponse) *structs.ClientHostVolumeConfig {
+func genVolConfig(req *cstructs.ClientHostVolumeCreateRequest, hostPath string) *structs.ClientHostVolumeConfig {
 	return &structs.ClientHostVolumeConfig{
 		Name: req.Name,
 		ID:   req.ID,
-		Path: resp.Path,
+		Path: hostPath,
 
 		// dynamic volumes, like CSI, have more robust `capabilities`,
 		// so we always set ReadOnly to false, and let the scheduler
